LPS-151639 Move _getDefaultLiferayHome() from PropsUtil to SystemProperties, prepare for next
